Fix an issue in simple-kernel-timer with mismatched linked symbols
Basic Kokkos Serial build and test failed for me on MacOS.
Symptom was that the simple kernel timer "currentEntry" symbol was null
during end of parallel for call back.
Tracked that down to that symbol having different addresses inside the
increment_counter function and the end of parallel_for callback.

I believe moving the increment_counter functions to not be inlined may help?

Or do we need accessor functions which are uniform?

At least this change did fix the issue for me.