Update kp_sampler_skip.cpp
Fixing  tool induced fences to always fence on device with DevID 0. 

Fencing with DevID will be a done in subsequent patch (where Pair object will be used in the hash table to capture the begin sample's information. 

Note that the pair/tuple object can capture other state information to store between the beginning of sampling event and ending of it.